This check-in is to remove all warning for both g++ 4.4 and g++ 4.2. Also this checkin makes the code directly compilable in both 32bit and 64 bit systems with their respective compilers. Changes the compilation to O3 and made some slight modifications to the M_incl file to have appropriate option for both systems. I am removing a .d and moc file from the svn that was messing up compilation for the 4.4 system. Adding touch.pl file for the 4.2 compiler and system to touch all .d files for compilation. Also all the xpm files being checked in are just to add a char* typecast to avoid warnings.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@186 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ Plum/ILoadable.h
@@ -107,7 +107,7 @@ namespace plum{
         string getPathName( const string filename )
         {
             char sep='/';
-            unsigned int pos=filename.rfind(sep);
+            size_t pos=filename.rfind(sep);
             if( pos==string::npos ) return string(); //empty
             string pathname(filename.begin(),filename.begin()+pos+1);
